% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_datasets.R
\name{combine_datasets}
\alias{combine_datasets}
\title{Combine data to be used in the applet into a single .RData file.}
\usage{
combine_datasets(shiny_setup, file_names, dataset)
}
\arguments{
\item{shiny_setup}{(required) The shiny_setup variable produced
during the \code{\link[=shiny_setup]{shiny_setup()}} function.}

\item{file_names}{(required) The file_names variable produced
during the \code{\link[=file_names]{file_names()}} function.}

\item{dataset}{(required) Whole brain dataset returned as an output of
the \code{\link[=forward_warp]{forward_warp()}} function.}
}
\description{
Produces a sunburst plot and data table, and stores them
with \emph{file_names} in an .RData file called full_data.R. This .RData
file is saved in the shiny output folder, and it is directly used by
the applet.
}
\details{
The new .RData file is automatically created in the RShiny
output folder.
}
