% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_names.R
\name{file_names}
\alias{file_names}
\title{Generate a list of file names for images output by SMART.}
\usage{
file_names(shiny_setup)
}
\arguments{
\item{shiny_setup}{(required) The shiny_setup variable produced
during the \code{\link[=shiny_setup]{shiny_setup()}} function.}
}
\value{
Returns \emph{file_names}, a list of length 4 containing
the following information: registration filenames, segmentation
filenames, forward warp filenames, and plate numbers.
}
\description{
Generates list of length 4 to store file names of
registration, segmentation, and forward warp images. These names
will be accessed downstream in the pipeline, including by future
functions and by the applet itself. This function requires that
the user has not modified the names of image files output by SMART.
}
