% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_selection.R
\name{file_selection}
\alias{file_selection}
\title{Isolate one segmentation image and one forward warp image
for each atlas plate.}
\usage{
file_selection(shiny_setup, file_names)
}
\arguments{
\item{shiny_setup}{(required) The shiny_setup variable produced
during the \code{\link[=shiny_setup]{shiny_setup()}} function.}

\item{file_names}{(required) The file_names variable produced
during the \code{\link[=file_names]{file_names()}} function.}
}
\description{
Copies one segmentation image and one forward warp
image per atlas plate from the SMART output folder (indicated
by the \emph{setup} variable) to the temporary directories in the
applet output folder. Because most whole brain datasets
contain thousands of segmentation and forward warp images, only
one of each, per atlas plate, will be displayed in the applet.
}
\details{
Files are automatically copied from the SMART output
folder to the temporary directories. The original image files
are kept intact.
}
