% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_conversion.R
\name{image_conversion}
\alias{image_conversion}
\title{Convert .tif files to .png files to conserve server space.}
\usage{
image_conversion(shiny_setup, file_names)
}
\arguments{
\item{shiny_setup}{(required) The shiny_setup variable produced
during the \code{\link[=shiny_setup]{shiny_setup()}} function.}

\item{file_names}{(required) The file_names variable produced
during the \code{\link[=file_names]{file_names()}} function.}
}
\description{
Converts registration, segmentation, and forward
warp images into .png files. One of each image type per atlas
plate is converted. The new .png files are stored in the shiny
output directories, and the temporary directories created by
\code{\link[=file_selection]{file_selection()}} are deleted.
}
\details{
Files are automatically copied to the temporary
directories. The original registration image files are
retained, but the .tif segmentation and forward warp images
are deleted along with their temporary directories. All
images in the SMART output folder are kept intact.
}
