% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartshiny.R
\docType{package}
\name{smartshiny}
\alias{smartshiny}
\alias{smartshiny-package}
\title{smartshiny: A package for building an RShiny app from SMART outut data}
\description{
The smartshiny package interfaces with the \link{SMART} package
by Michelle Jin, which is used in conjunction with the \link{wholebrain} package
by Daniel Furth to process whole brain imaging datasets. This package is
used to produce an RShiny app from the output data from the \link{SMART} package.
The applet production pipeline consists of four parts, as outlined in the
online tutorial. The first two parts contain new functions, as described
below.
}
\section{Part 1. Data collection}{

\enumerate{
\item \code{\link[=shiny_setup]{shiny_setup()}} Generate a list of parameters and file directory for the applet and related files.
}
}

\section{Part 2. Data processing}{

\enumerate{
\item \code{\link[=file_names]{file_names()}} Generate a list of file names for images output by SMART.
\item \code{\link[=file_selection]{file_selection()}} Isolate one segmentation image and one forward warp image for each atlas plate.
\item \code{\link[=image_conversion]{image_conversion()}} Convert .tif files to .png to conserve server space.
\item \code{\link[=combine_data]{combine_data()}} Combine data to be used in the applet into a single .RData file.
}
}

